batches .long() datatype

+++ model/training/dataset.py
@@ -12,6 +12,6 @@ class TokenDataset(Dataset):
         return len(self.tokens) - self.block_size
     
     def __getitem__(self, idx):
-        x = torch.from_numpy(self.tokens[idx : idx + self.block_size])
-        y = torch.from_numpy(self.tokens[idx + 1 : idx + self.block_size + 1])
+        x = torch.from_numpy(self.tokens[idx : idx + self.block_size]).long()
+        y = torch.from_numpy(self.tokens[idx + 1 : idx + self.block_size + 1]).long()
         return x, y
